Cart And Stripe Modification

+++ app/Http/Controllers/ItemController.php
@@ -7,6 +7,7 @@ use App\Models\Category;
 use App\Models\Item;
 use Carbon\Carbon;
 use Illuminate\Http\RedirectResponse;
+use Illuminate\Support\Facades\Storage;
 use Illuminate\View\View;
 
 class ItemController extends Controller
@@ -14,16 +15,11 @@ class ItemController extends Controller
 
     public function index(): View
     {
-        $user = auth()->user();
+        $this->authorize('viewAny', \auth()->user());
 
-        $items = Item::with('categories')
-            ->whereHas('categories', function ($query) use ($user) {
-                $query->whereHas('restaurant', function ($query) use ($user) {
-                    $query->where('user_id', $user->id);
-                });
-            })
-            ->get();
+        $items = Item::where('restaurant_id',auth()->user()->restaurant->id)->get();
         return view('item.index', compact('items'));
+
     }
 
     /**
@@ -31,19 +27,24 @@ class ItemController extends Controller
      */
     public function create(): View
     {
-        $categories = auth()->user()->restaurant->categories->whereNotNull('parent_id');
+        $this->authorize('create', \auth()->user());
 
+        $categories = auth()->user()->restaurant->categories;
         return view('item.create', compact('categories'));
     }
 
     public function store(ItemRequest $request): RedirectResponse
     {
+        $this->authorize('create', \auth()->user());
         $validatedData = $request->validated();
         $validatedData['category_id'] = $request->category_id;
 
-        if ($request->hasFile('image') && $request->image != null) {
+        $validatedData['restaurant_id'] = auth()->user()->restaurant->id;
+
+        if ($request->hasFile('image')) {
+
             $imageName = $request->image->getClientOriginalName();
-            $request->image->storeAs('public/images', $imageName);
+            $request->file('image')->store('images', 'public');
 
             $validatedData['image'] = $imageName;
         }
@@ -53,7 +54,7 @@ class ItemController extends Controller
 
         $item->categories()->attach($category, ['created_at' => Carbon::now(), 'updated_at' => Carbon::now()]);
 
-        return redirect()->route('items.index')->with('status', 'Item Created Successfully');
+        return redirect()->route('items.index')->with('status', 'تمت إضافة عنصر بنجاح');
     }
 
     public function show(Item $item): View
@@ -63,15 +64,30 @@ class ItemController extends Controller
 
     public function edit(Item $item): View
     {
-        return view('item.edit', compact('item'));
+        $this->authorize('update', \auth()->user());
+
+        $categories = auth()->user()->restaurant->categories->whereNotNull('parent_id');
+        return view('item.edit', compact('item','categories'));
     }
 
     public function update(ItemRequest $request, Item $item): RedirectResponse
     {
+        $this->authorize('update', \auth()->user());
+        $old_image = $item->image;
+
         $validatedDate = $request->validated();
+        if ($request->hasFile('image')){
+            $image_name = $request->input('image')->getClientOriginalName();
+
+            $request->file('image')->store('images', 'public');
+            $item->image = $image_name;
+        }
 
         $item->update($validatedDate);
-        return redirect()->route('items.index');
+        if ($old_image && $validatedDate['image']){
+            Storage::disk('public')->delete($old_image);
+        }
+        return redirect()->route('items.index')->with('status','تم التحديث بنجاح');
     }
 
     /**
@@ -79,7 +95,9 @@ class ItemController extends Controller
      */
     public function destroy(Item $item): RedirectResponse
     {
+        $this->authorize('delete', \auth()->user());
+
         $item->delete();
-        return redirect()->back()->with('status', 'Item Deleted Successfully');
+        return redirect()->back()->with('status', 'تم الحذف بنجاح');
     }
 }
